@@ -1,74 +1,76 @@
-module github.com/containrrr/watchtower
+module github.com/Marrrrrrrrry/watchtower
 
-go 1.20
+go 1.25
 
 require (
-	github.com/containrrr/shoutrrr v0.8.0
-	github.com/distribution/reference v0.5.0
-	github.com/docker/cli v24.0.7+incompatible
-	github.com/docker/docker v24.0.7+incompatible
-	github.com/docker/go-connections v0.4.0
-	github.com/onsi/ginkgo v1.16.5
-	github.com/onsi/gomega v1.30.0
-	github.com/prometheus/client_golang v1.18.0
-	github.com/robfig/cron v1.2.0
-	github.com/sirupsen/logrus v1.9.3
-	github.com/spf13/cobra v1.8.0
-	github.com/spf13/pflag v1.0.5
-	github.com/spf13/viper v1.18.2
-	github.com/stretchr/testify v1.8.4
-	golang.org/x/net v0.19.0
+    github.com/containrrr/shoutrrr v0.8.0
+    github.com/distribution/reference v0.6.0
+    github.com/docker/cli v28.5.2+incompatible
+    github.com/docker/docker v28.5.2+incompatible
+    github.com/docker/go-connections v0.6.0
+    github.com/onsi/ginkgo v1.16.5
+    github.com/onsi/gomega v1.38.2
+    github.com/prometheus/client_golang v1.23.2
+    github.com/robfig/cron v1.2.0
+    github.com/sirupsen/logrus v1.9.3
+    github.com/spf13/cobra v1.10.1
+    github.com/spf13/pflag v1.0.5
+    github.com/spf13/viper v1.21.0
+    github.com/stretchr/testify v1.11.1
+    golang.org/x/net v0.38.0
 )
 
 require github.com/matttproud/golang_protobuf_extensions/v2 v2.0.0 // indirect
 
+replace github.com/containrrr/watchtower => .
+
 require (
-	github.com/Azure/go-ansiterm v0.0.0-20210617225240-d185dfc1b5a1 // indirect
-	github.com/Microsoft/go-winio v0.4.17 // indirect
-	github.com/beorn7/perks v1.0.1 // indirect
-	github.com/cespare/xxhash/v2 v2.2.0 // indirect
-	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
-	github.com/docker/distribution v2.8.3+incompatible // indirect
-	github.com/docker/docker-credential-helpers v0.6.1 // indirect
-	github.com/docker/go-units v0.4.0 // indirect
-	github.com/fatih/color v1.15.0 // indirect
-	github.com/fsnotify/fsnotify v1.7.0 // indirect
-	github.com/gogo/protobuf v1.3.2 // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/google/go-cmp v0.6.0 // indirect
-	github.com/hashicorp/hcl v1.0.0 // indirect
-	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/magiconair/properties v1.8.7 // indirect
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.17 // indirect
-	github.com/mitchellh/mapstructure v1.5.0 // indirect
-	github.com/moby/term v0.0.0-20210619224110-3f7ff695adc6 // indirect
-	github.com/morikuni/aec v0.0.0-20170113033406-39771216ff4c // indirect
-	github.com/nxadm/tail v1.4.8 // indirect
-	github.com/opencontainers/go-digest v1.0.0 // indirect
-	github.com/opencontainers/image-spec v1.0.2 // indirect
-	github.com/pelletier/go-toml/v2 v2.1.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
-	github.com/prometheus/client_model v0.5.0 // indirect
-	github.com/prometheus/common v0.45.0 // indirect
-	github.com/prometheus/procfs v0.12.0 // indirect
-	github.com/sagikazarmark/locafero v0.4.0 // indirect
-	github.com/sagikazarmark/slog-shim v0.1.0 // indirect
-	github.com/sourcegraph/conc v0.3.0 // indirect
-	github.com/spf13/afero v1.11.0 // indirect
-	github.com/spf13/cast v1.6.0 // indirect
-	github.com/stretchr/objx v0.5.0 // indirect
-	github.com/subosito/gotenv v1.6.0 // indirect
-	go.uber.org/atomic v1.9.0 // indirect
-	go.uber.org/multierr v1.9.0 // indirect
-	golang.org/x/exp v0.0.0-20230905200255-921286631fa9 // indirect
-	golang.org/x/sys v0.15.0 // indirect
-	golang.org/x/text v0.14.0
-	golang.org/x/time v0.5.0 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
-	gopkg.in/ini.v1 v1.67.0 // indirect
-	gopkg.in/tomb.v1 v1.0.0-20141024135613-dd632973f1e7 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-	gotest.tools/v3 v3.0.3 // indirect
+    github.com/Azure/go-ansiterm v0.0.0-20210617225240-d185dfc1b5a1 // indirect
+    github.com/Microsoft/go-winio v0.6.2 // indirect
+    github.com/beorn7/perks v1.0.1 // indirect
+    github.com/cespare/xxhash/v2 v2.3.0 // indirect
+    github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
+    github.com/docker/distribution v2.8.3+incompatible // indirect
+    github.com/docker/docker-credential-helpers v0.8.0 // indirect
+    github.com/docker/go-units v0.5.0 // indirect
+    github.com/fatih/color v1.16.0 // indirect
+    github.com/fsnotify/fsnotify v1.9.0 // indirect
+    github.com/gogo/protobuf v1.3.2 // indirect
+    github.com/golang/protobuf v1.5.4 // indirect
+    github.com/google/go-cmp v0.6.0 // indirect
+    github.com/hashicorp/hcl v1.0.0 // indirect
+    github.com/inconshreveable/mousetrap v1.1.0 // indirect
+    github.com/magiconair/properties v1.8.7 // indirect
+    github.com/mattn/go-colorable v0.1.13 // indirect
+    github.com/mattn/go-isatty v0.0.20 // indirect
+    github.com/mitchellh/mapstructure v1.5.0 // indirect
+    github.com/moby/term v0.5.0 // indirect
+    github.com/morikuni/aec v0.0.0-20170113033406-39771216ff4c // indirect
+    github.com/nxadm/tail v1.4.8 // indirect
+    github.com/opencontainers/go-digest v1.0.0 // indirect
+    github.com/opencontainers/image-spec v1.1.0 // indirect
+    github.com/pelletier/go-toml/v2 v2.2.4 // indirect
+    github.com/pkg/errors v0.9.1 // indirect
+    github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
+    github.com/prometheus/client_model v0.6.2 // indirect
+    github.com/prometheus/common v0.66.1 // indirect
+    github.com/prometheus/procfs v0.12.0 // indirect
+    github.com/sagikazarmark/locafero v0.9.0 // indirect
+    github.com/sagikazarmark/slog-shim v0.1.0 // indirect
+    github.com/sourcegraph/conc v0.3.0 // indirect
+    github.com/spf13/afero v1.11.0 // indirect
+    github.com/spf13/cast v1.9.2 // indirect
+    github.com/stretchr/objx v0.5.2 // indirect
+    github.com/subosito/gotenv v1.6.0 // indirect
+    go.uber.org/atomic v1.11.0 // indirect
+    go.uber.org/multierr v1.11.0 // indirect
+    golang.org/x/exp v0.0.0-20230905200255-921286631fa9 // indirect
+    golang.org/x/sys v0.38.0 // indirect
+    golang.org/x/text v0.14.0
+    golang.org/x/time v0.5.0 // indirect
+    google.golang.org/protobuf v1.36.6 // indirect
+    gopkg.in/ini.v1 v1.67.0 // indirect
+    gopkg.in/tomb.v1 v1.0.0-20141024135613-dd632973f1e7 // indirect
+    gopkg.in/yaml.v3 v3.0.1 // indirect
+    gotest.tools/v3 v3.0.3 // indirect
 )
