@@ -5,10 +5,10 @@ import (
 	"sort"
 	"time"
 
-	"github.com/containrrr/watchtower/pkg/container"
-	"github.com/containrrr/watchtower/pkg/filters"
-	"github.com/containrrr/watchtower/pkg/sorter"
-	"github.com/containrrr/watchtower/pkg/types"
+	"github.com/Marrrrrrrrry/watchtower/pkg/container"
+	"github.com/Marrrrrrrrry/watchtower/pkg/filters"
+	"github.com/Marrrrrrrrry/watchtower/pkg/sorter"
+	"github.com/Marrrrrrrrry/watchtower/pkg/types"
 
 	log "github.com/sirupsen/logrus"
 )
