@@ -1,21 +1,22 @@
 package mocks
 
 import (
-	"encoding/json"
-	"fmt"
-	"github.com/onsi/ginkgo"
-	"net/http"
-	"net/url"
-	"os"
-	"path/filepath"
-	"strings"
+    "encoding/json"
+    "fmt"
+    "github.com/onsi/ginkgo"
+    "net/http"
+    "net/url"
+    "os"
+    "path/filepath"
+    "strings"
 
-	t "github.com/containrrr/watchtower/pkg/types"
+    t "github.com/Marrrrrrrrry/watchtower/pkg/types"
 
-	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/api/types/filters"
-	O "github.com/onsi/gomega"
-	"github.com/onsi/gomega/ghttp"
+    types "github.com/docker/docker/api/types"
+    imageTypes "github.com/docker/docker/api/types/image"
+    "github.com/docker/docker/api/types/filters"
+    O "github.com/onsi/gomega"
+    "github.com/onsi/gomega/ghttp"
 )
 
 func getMockJSONFile(relPath string) ([]byte, error) {
@@ -262,14 +263,14 @@ func RemoveImageHandler(imagesWithParents map[string][]string) http.HandlerFunc 
 			image := parts[len(parts)-1]
 
 			if parents, found := imagesWithParents[image]; found {
-				items := []types.ImageDeleteResponseItem{
-					{Untagged: image},
-					{Deleted: image},
-				}
-				for _, parent := range parents {
-					items = append(items, types.ImageDeleteResponseItem{Deleted: parent})
-				}
-				ghttp.RespondWithJSONEncoded(http.StatusOK, items)(w, r)
+                items := []imageTypes.DeleteResponse{
+                    {Untagged: image},
+                    {Deleted: image},
+                }
+                for _, parent := range parents {
+                    items = append(items, imageTypes.DeleteResponse{Deleted: parent})
+                }
+                ghttp.RespondWithJSONEncoded(http.StatusOK, items)(w, r)
 			} else {
 				ghttp.RespondWithJSONEncoded(http.StatusNotFound, struct{ message string }{
 					message: "Something went wrong.",
