@@ -6,7 +6,7 @@ import (
 	"github.com/spf13/cobra"
 
 	shoutrrrSmtp "github.com/containrrr/shoutrrr/pkg/services/smtp"
-	t "github.com/containrrr/watchtower/pkg/types"
+	t "github.com/Marrrrrrrrry/watchtower/pkg/types"
 	log "github.com/sirupsen/logrus"
 )
 
